[chore] #61 BigDecimal 변경

+++ module-domain/domain-products/src/main/java/com/chaing/domain/products/dto/request/ProductUpdateRequest.java
@@ -2,15 +2,16 @@ package com.chaing.domain.products.dto.request;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
 @Builder
 public record ProductUpdateRequest(
         String name,                // 상품명
-        Integer price,              // 판매가
-        Integer originalPrice,      // 원가
-        Integer supplyPrice,        // 공급가
+        BigDecimal price,              // 판매가
+        BigDecimal originalPrice,      // 원가
+        BigDecimal supplyPrice,        // 공급가
         String status,              // 상태
         Integer baseSafeStock,      // 기준 안전재고
         Integer kcal,               // 칼로리
